+++ system/modules/isotope/IsotopeFrontend.php
@@ -676,9 +676,8 @@ $endScript";
 				// Store current value in the session
 				elseif ($objWidget->submitInput())
 				{
-					$objForm->arrFormData[$objFields->name]	= $objWidget->value;
-					$_SESSION['FORM_DATA'][$objFields->name]		= $objWidget->value;
-
+					$objForm->arrFormData[$objFields->name] = $objWidget->value;
+					$_SESSION['FORM_DATA'][$objFields->name] = $objWidget->value;
 				}
 
 				// Store file uploads
@@ -1316,6 +1315,22 @@ $endScript";
 		$arrIsotopeProductPages = array_unique($arrIsotopeProductPages);
 		
 		return array_merge($arrPages, $arrIsotopeProductPages);
+	}
+	
+	
+	/**
+	 * save_callback for upload widget to store $_FILES data into the product
+	 * @param mixed
+	 * @param IsotopeProduct
+	 */
+	public function saveUpload($varValue, IsotopeProduct $objProduct, Widget $objWidget)
+	{
+		if (is_array($_SESSION['FILES'][$objWidget->name]) && $_SESSION['FILES'][$objWidget->name]['uploaded'] == '1' && $_SESSION['FILES'][$objWidget->name]['error'] == 0)
+		{
+			return $_SESSION['FILES'][$objWidget->name]['name'];
+		}
+
+		return $varValue;
 	}