+++ system/modules/isotope/dca/tl_iso_mail.php
@@ -191,7 +191,7 @@ $GLOBALS['TL_DCA']['tl_iso_mail'] = array
 			'inputType'               => 'select',
 			'default'                 => 'mail_default',
 			'options'                 => IsotopeBackend::getTemplates('mail_'),
-			'eval'                    => array('mandatory'=>true, 'tl_class'=>'w50'),
+			'eval'                    => array('mandatory'=>true, 'tl_class'=>'w50', 'chosen'=>true)
 		),
 		'priority' => array
 		(
@@ -216,7 +216,7 @@ $GLOBALS['TL_DCA']['tl_iso_mail'] = array
 			'exclude'                 => true,
 			'inputType'				  => 'select',
 			'options'				  => IsotopeBackend::getTemplates('iso_invoice'),
-			'eval'					  => array('includeBlankOption'=>true, 'tl_class'=>'w50'),
+			'eval'					  => array('includeBlankOption'=>true, 'tl_class'=>'w50', 'chosen'=>true)
 		),
 		'documentTitle'		=> array
 		(