+++ system/modules/isotope/languages/de/tl_iso_attributes.php
@@ -127,4 +127,9 @@ $GLOBALS['TL_LANG']['tl_iso_attributes']['options_legend'] = 'Optionen';
 $GLOBALS['TL_LANG']['tl_iso_attributes']['config_legend'] = 'Attribut-Konfiguration';
 $GLOBALS['TL_LANG']['tl_iso_attributes']['validation_legend'] = 'Eingabeprüfung';
 $GLOBALS['TL_LANG']['tl_iso_attributes']['search_filters_legend'] = 'Suche- & Filter-Einstellungen';
+$GLOBALS['TL_LANG']['tl_iso_attributes']['storeFile']      = array('Hochgeladene Dateien speichern', 'Die hochgeladenen Dateien in einen Ordner auf dem Server verschieben.');
+$GLOBALS['TL_LANG']['tl_iso_attributes']['uploadFolder']   = array('Zielverzeichnis', 'Bitte wählen Sie das Zielverzeichnis aus der Dateiübersicht.');
+$GLOBALS['TL_LANG']['tl_iso_attributes']['useHomeDir']     = array('Benutzerverzeichnis verwenden', 'Die Datei im Benutzerverzeichnis speichern, wenn sich ein Benutzer angemeldet hat.');
+$GLOBALS['TL_LANG']['tl_iso_attributes']['doNotOverwrite'] = array('Bestehende Dateien erhalten', 'Der neuen Datei ein numerisches Suffix hinzufügen, wenn der Dateiname bereits existiert.');
+$GLOBALS['TL_LANG']['tl_iso_attributes']['store_legend']   = 'Datei speichern';